+++ src/main/java/me/hugmanrique/cartage/compression/RLUnCompDecompressor.java
@@ -28,9 +28,9 @@ public final class RLUnCompDecompressor implements Decompressor {
 
   // TODO error checking
   @Override
-  public byte[] decompress(final Cartridge cartridge, final int start) {
+  public byte[] decompress(final Cartridge cartridge, final long start) {
     // TODO Check cartridge.remaining >= 4
-    int offset = start;
+    long offset = start;
     final int header = cartridge.getInt(offset);
     offset += 4;
     if ((header >>> 24) != MAGIC_NUMBER) {
@@ -42,7 +42,7 @@ public final class RLUnCompDecompressor implements Decompressor {
     if (length == 0) {
       return result;
     }
-    final int end = start + length;
+    final long end = start + length;
     byte runLength;
 
     do {
